requirements.txt
click==7.1.1
cloudpickle==1.2.2
cycler==0.10.0
Flask==1.1.2
future==0.18.0
gym==0.15.3
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.3
opencv-python==4.1.1.26
Pillow==7.1.1
pyglet==1.3.2
pyparsing==2.4.4
python-dateutil==2.8.1
scipy==1.3.1
six==1.13.0
torch==1.3.0
Werkzeug==1.0.1
